Extract project geometry queries
Includes additional helper for combining multile feature collections;
