+++ app/models/project.js
@@ -2,11 +2,7 @@ import DS from 'ember-data';
 import { attr, hasMany } from '@ember-decorators/data';
 import { computed } from '@ember-decorators/object';
 import turfBbox from '@turf/bbox';
-import turfUnion from '@turf/union';
-import turfBuffer from '@turf/buffer';
-import turfDifference from '@turf/difference';
 import { camelize } from '@ember/string';
-import carto from 'cartobox-promises-utility/utils/carto';
 import {
   type,
   arrayOf,
@@ -15,10 +11,13 @@ import {
   optional,
   oneOf,
 } from '@ember-decorators/argument/type';
-import isEmpty from '../utils/is-empty';
-import config from '../config/environment';
+import intersectingZoningQuery from 'labs-applicant-maps/utils/queries/intersecting-zoning-query';
+import proposedCommercialOverlaysQuery from 'labs-applicant-maps/utils/queries/proposed-commercial-overlays-query';
+import proposedSpecialDistrictsQuery from 'labs-applicant-maps/utils/queries/proposed-special-districts-query';
+import rezoningAreaQuery from 'labs-applicant-maps/utils/queries/rezoning-area-query';
+import isEmpty from 'labs-applicant-maps/utils/is-empty';
+import config from 'labs-applicant-maps/config/environment';
 
-const bufferMeters = 500;
 const { mapTypes } = config;
 const { Model } = DS;
 
@@ -63,128 +62,8 @@ export const FeatureCollection = shapeOf({
   ),
 });
 
-export const INTERSECTING_ZONING_QUERY = async (developmentSite) => {
-  if (developmentSite) {
-    // Get zoning districts
-    const zoningQuery = `
-      WITH buffer as (
-        SELECT ST_SetSRID(
-          ST_Buffer(
-            ST_GeomFromGeoJSON('${JSON.stringify(developmentSite)}')::geography,
-            ${bufferMeters}
-          ),
-        4326)::geometry AS the_geom
-      )
-      SELECT ST_Intersection(zoning.the_geom, buffer.the_geom) AS the_geom, zonedist AS label, cartodb_id AS id
-      FROM planninglabs.zoning_districts_v201809 zoning, buffer
-      WHERE ST_Intersects(zoning.the_geom,buffer.the_geom)
-    `;
 
-    const clippedZoningDistricts = await new carto.SQL(zoningQuery, 'geojson');
-
-    // add an id to the top level of each feature object, for use by mapbox-gl-draw
-    const { features } = clippedZoningDistricts;
-    clippedZoningDistricts.features = features.map((feature) => {
-      feature.id = feature.properties.id;
-      return feature;
-    });
-
-    return clippedZoningDistricts;
-  }
-
-  return null;
-};
-
-export const PROPOSED_COMMERCIAL_OVERLAYS_QUERY = (developmentSite) => {
-  if (developmentSite) {
-    // Get commercial overlays
-    const commercialOverlaysQuery = `
-      WITH buffer as (
-        SELECT ST_SetSRID(
-          ST_Buffer(
-            ST_GeomFromGeoJSON('${JSON.stringify(developmentSite)}')::geography,
-            ${bufferMeters}
-          ),
-        4326)::geometry AS the_geom
-      )
-      SELECT ST_Intersection(co.the_geom, buffer.the_geom) AS the_geom, overlay AS label
-      FROM planninglabs.commercial_overlays_v201809 co, buffer
-      WHERE ST_Intersects(co.the_geom,buffer.the_geom)
-    `;
-
-    return new carto.SQL(commercialOverlaysQuery, 'geojson');
-  }
-
-  return null;
-};
-
-export const PROPOSE_SPECIAL_DISTRICTS_QUERY = (developmentSite) => {
-  if (developmentSite) {
-    // Get special purpose districts
-    const specialPurposeDistrictsQuery = `
-      WITH buffer as (
-        SELECT ST_SetSRID(
-          ST_Buffer(
-            ST_GeomFromGeoJSON('${JSON.stringify(developmentSite)}')::geography,
-            ${bufferMeters}
-          ),
-        4326)::geometry AS the_geom
-      )
-      SELECT ST_Intersection(spd.the_geom, buffer.the_geom) AS the_geom, sdname AS label
-      FROM planninglabs.special_purpose_districts_v201809 spd, buffer
-      WHERE ST_Intersects(spd.the_geom,buffer.the_geom)
-    `;
-
-    return new carto.SQL(specialPurposeDistrictsQuery, 'geojson');
-  }
-
-  return null;
-};
-
-export const REZONING_AREA_QUERY = (currentZoning, proposedZoning) => {
-  // create an empty FeatureCollection to hold the difference sections
-  const differenceFC = {
-    type: 'FeatureCollection',
-    features: [],
   };
-
-  // flag differences
-  proposedZoning.features.forEach((feature) => {
-    const { id } = feature;
-    const correspondingCurrentZoningFeature = currentZoning.features.filter(d => d.id === id)[0];
-
-    // if feature exists in currentZoning, compare the geometries
-    if (correspondingCurrentZoningFeature) {
-      // get the difference
-      const difference = turfDifference(correspondingCurrentZoningFeature, feature);
-      if (difference) differenceFC.features.push(difference);
-
-      // get the inverse difference (reverse the order of the polygons)
-      const inverseDifference = turfDifference(feature, correspondingCurrentZoningFeature);
-      if (inverseDifference) differenceFC.features.push(inverseDifference);
-    } else {
-      differenceFC.features.push(feature);
-    }
-  });
-
-  // union together all difference features
-  if (differenceFC.features.length > 0) {
-    const differenceUnion = differenceFC.features
-      .reduce((union, { geometry }) => {
-        if (union === null) {
-          union = geometry;
-        } else {
-          union = turfUnion(union, geometry);
-        }
-
-        const buffered = turfBuffer(union, -0.0005);
-        return buffered;
-      }, null);
-
-    return differenceUnion;
-  }
-
-  return null;
 };
 
 const trueOrNull = property => property === true || property === null;
@@ -242,7 +121,7 @@ export default class extends Model {
 
   async setDefaultUnderlyingZoning() {
     const developmentSite = this.get('developmentSite');
-    const result = await INTERSECTING_ZONING_QUERY(developmentSite);
+    const result = await intersectingZoningQuery(developmentSite);
 
     this.set('underlyingZoning', result);
   }
@@ -254,7 +133,7 @@ export default class extends Model {
 
   async setDefaultCommercialOverlays() {
     const developmentSite = this.get('developmentSite');
-    const result = await PROPOSED_COMMERCIAL_OVERLAYS_QUERY(developmentSite);
+    const result = await proposedCommercialOverlaysQuery(developmentSite);
 
     this.set('commercialOverlays', result);
   }
@@ -266,7 +145,7 @@ export default class extends Model {
 
   async setDefaultSpecialPurposeDistricts() {
     const developmentSite = this.get('developmentSite');
-    const result = await PROPOSE_SPECIAL_DISTRICTS_QUERY(developmentSite);
+    const result = await proposedSpecialDistrictsQuery(developmentSite);
 
     this.set('specialPurposeDistricts', result);
   }
@@ -278,8 +157,8 @@ export default class extends Model {
   async setRezoningArea() {
     const proposedZoning = this.get('underlyingZoning');
     const developmentSite = this.get('developmentSite');
-    const currentZoning = await PROPOSE_SPECIAL_DISTRICTS_QUERY(developmentSite);
-    const result = await REZONING_AREA_QUERY(currentZoning, proposedZoning);
+    const currentZoning = await proposedSpecialDistrictsQuery(developmentSite);
+    const result = await rezoningAreaQuery(currentZoning, proposedZoning);
 
     this.set('rezoningArea', result);
   }
