+++ app/utils/combine-to-multi-polygon.js
@@ -1,0 +1,6 @@
+import turfUnion from '@turf/union';
+
+export default function unionFromFeatureCollection({ features }) {
+  console.log(features);
+  return turfUnion(...features);
+}
