+++ config/environment.js
@@ -24,6 +24,9 @@ module.exports = function (environment) {
     // reusable list of named map types
     mapTypes: ['area-maps', 'tax-maps', 'zoning-change-maps', 'zoning-section-maps'],
 
+    // buffer in meters for queries
+    bufferMeters: 500,
+
     APP: {
       // Here you can pass flags/options to your application instance
       // when it is created
