+++ package.json
@@ -84,6 +84,7 @@
     "@turf/bbox": "^6.0.1",
     "@turf/buffer": "^5.1.5",
     "@turf/centroid": "^6.0.2",
+    "@turf/combine": "6.0.1",
     "@turf/difference": "6.0.2",
     "@turf/simplify": "^5.1.5",
     "@turf/union": "^6.0.3",
