+++ tests/unit/utils/queries/intersecting-zoning-query-test.js
@@ -1,0 +1,11 @@
+import queriesIntersectingZoningQuery from 'labs-applicant-maps/utils/queries/intersecting-zoning-query';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | queries/intersecting-zoning-query', function(hooks) {
+
+  // Replace this with your real tests.
+  test('it works', function(assert) {
+    let result = queriesIntersectingZoningQuery();
+    assert.ok(result);
+  });
+});
