+++ tests/unit/utils/queries/proposed-commercial-overlays-query-test.js
@@ -1,0 +1,11 @@
+import queriesProposedCommercialOverlaysQuery from 'labs-applicant-maps/utils/queries/proposed-commercial-overlays-query';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | queries/proposed-commercial-overlays-query', function(hooks) {
+
+  // Replace this with your real tests.
+  test('it works', function(assert) {
+    let result = queriesProposedCommercialOverlaysQuery();
+    assert.ok(result);
+  });
+});
