+++ tests/unit/utils/queries/proposed-special-districts-query-test.js
@@ -1,0 +1,11 @@
+import queriesProposedSpecialDistrictsQuery from 'labs-applicant-maps/utils/queries/proposed-special-districts-query';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | queries/proposed-special-districts-query', function(hooks) {
+
+  // Replace this with your real tests.
+  test('it works', function(assert) {
+    let result = queriesProposedSpecialDistrictsQuery();
+    assert.ok(result);
+  });
+});
