+++ tests/unit/utils/queries/rezoning-area-query-test.js
@@ -1,0 +1,11 @@
+import queriesRezoningAreaQuery from 'labs-applicant-maps/utils/queries/rezoning-area-query';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | queries/rezoning-area-query', function(hooks) {
+
+  // Replace this with your real tests.
+  test('it works', function(assert) {
+    let result = queriesRezoningAreaQuery();
+    assert.ok(result);
+  });
+});
