+++ tests/unit/utils/unify-polygons-test.js
@@ -1,0 +1,11 @@
+import unifyPolygons from 'labs-applicant-maps/utils/unify-polygons';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | unify-polygons', function(hooks) {
+
+  // Replace this with your real tests.
+  test('it works', function(assert) {
+    let result = unifyPolygons();
+    assert.ok(result);
+  });
+});
